+++ common/src/main/java/com/selfdot/cobblemontrainers/util/DataKeys.java
@@ -11,6 +11,7 @@ public class DataKeys {
     public static final String POKEMON_ABILITY = "ability";
     public static final String POKEMON_MOVESET = "moveset";
     public static final String POKEMON_SHINY = "shiny";
+    public static final String POKEMON_HELD_ITEM = "heldItem";
     public static final String TRAINER_NAME = "name";
     public static final String TRAINER_TEAM = "team";
     public static final String TRAINER_MONEY_REWARD = "moneyReward";